Add security-key field to Ru/Sa save-structure.
Ruby/Sapphire games do not protect important data such as money, item
quantities etc. with an XOR-encryption like Em/FR/LG do, but adding support for
it is useful because of the following:

1. The field specified is guaranteed to always be 0 for Ru/Sa, thus it can be
applied as an en-/decryption key without changing data that does not need to be
en-/decrypted for Ru/Sa games.

2. Methods accessing encrypted data can be simplified to always apply a key,
even when none is needed. This removes the need to handle Ru/Sa games
seperately.