Correct call to malloc().
Allocate memory for all 14 save sections because all are going to be copied,
even if not every byte of them will be used by the corresponding unpacked save
structure.

PS: We were allocating memory for a pointer before. 8 bytes != 14*0xff4 bytes.
We may have been a bit off with the size there. Thanks valgrind for noticing!