Remove magic numbers from code.
Every constant now gets the #define that it deserves, reducing the chances of
accidentally introducing bugs when fiddling with indeces, block-counts and
strings-sizes. Also makes the code more readable; named constants convey more
information on how, or why, a number is used in a certain context.

Consider:

    for (size_t i = 0; i < 6; i++) {
        //do stuff...
    }

and:

    for (size_t i = 0; i < SAVE_PARTY_SLOTS; i++) {
        //do stuff...
    }

One of these two cases does not need any more context to express that it
iterates over every party slot of a given save.